+++ CHANGES.md
@@ -7,6 +7,18 @@ Version 0.2.0
 To be released.
 
 
+Version 0.1.1
+-------------
+
+Released on August 21, 2025.
+
+ -  Fixed a bug where `object()` parsers containing only Boolean flags would
+    fail when no arguments were provided, instead of defaulting the flags to
+    `false`. [[#6]]
+
+[#6]: https://github.com/dahlia/optique/issues/6
+
+
 Version 0.1.0
 -------------